Updated readme to include instructions on how to use new functionality.
New features:

-Creating a user automatically creates an ACL for the user along with
ACEs that allow a user read, write, and delete permission for thier own
User object.

-Deleting a user automatically removes any ACEs and Authorities
associated with that user. It also sets any ACLs owned by the deleted
user to have owner_sid as NULL.

-GenericAclController is available for any new resource managing
services to use.  It contains methods for maintaining ACLs and ACEs of
any object that implements interface IAclObject.

Changes:
-deleteUser now requires that the request contain a JSON of the user to
be deleted and the JSON must include id set to not null.

-reworked the mysql tables so that they handle user deletion gracefully